fixed deleteController logic

+++ modules/controllers/api/v1/adminController/role/destroyController.js
@@ -1,27 +1,28 @@
 import User from "../../../../../models/user-model.js";
-import { InitializeController } from "./initializeController.js";
+import {
+    InitializeController
+} from "./initializeController.js";
 
 
-export default new (class destroyController extends InitializeController{
+export default new(class destroyController extends InitializeController {
 
-    async deleteRole (req, res, next)  {
+    async deleteRole(req, res, next) {
         try {
             const roleId = req.params.roleId;
+            if (!roleId) return this.abort(res, 400, null, "id is undefined")
             const role = await this.model.Role.findByIdAndDelete(roleId);
-            if(!role) return this.abort(res , 404 , null , "role not found")
+            if (!role) return this.abort(res, 404, null, "role not found")
             const user = await User.findById(role.userRef);
-            const newArr = user.role.filter(e => e !== role.role );
-            if(newArr.length === 0) user.remove();
-            else{
+            const newArr = user.role.filter(e => e !== role.role);
+            if (newArr.length === 0) user.remove();
+            else {
                 user.role = newArr;
                 await user.save()
             }
-            this.helper.response(res, null , null , 200 ,{
+            this.helper.response(res, null, null, 200, {
                 success: true,
                 message: 'Role has been deleted'
             })
-            
-            // getGrants();
         } catch (error) {
             next(error)
         }
